strip(tier-1): delete crates/warp_web_event_bus
The web event bus existed to push events from a wasm-embedded warp
into the host JavaScript page. This fork is a native macOS terminal,
so the bus is unused.

- crates/warp_web_event_bus/ deleted
- removed from [workspace.dependencies] and from warp_logging + app
- single call site in warp_logging/src/wasm.rs replaced with a no-op
- re-export in app/src/platform/wasm.rs replaced with a local stub
  enum + no-op fn so any future wasm-target build still compiles

Both wasm.rs files are cfg(target_family = "wasm") only and never
participate in the macOS build.
