+++ crates/warp_logging/Cargo.toml
@@ -27,7 +27,6 @@ console_error_panic_hook = { version = "0.1.6" }
 js-sys = "0.3.63"
 wasm-bindgen.workspace = true
 web-sys.workspace = true
-warp_web_event_bus.workspace = true
 
 [features]
 crash_reporting = ["dep:sentry", "dep:sentry-log", "warp_core/crash_reporting"]
