--- crates/warp_web_event_bus/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "warp_web_event_bus"
-edition = "2024"
-publish.workspace = true
-license.workspace = true
-
-[lib]
-name = "warp_web_event_bus"
-path = "src/lib.rs"
-
-[dependencies]
-serde = { workspace = true }
-wasm-bindgen = { workspace = true }
-serde-wasm-bindgen = "0.6.5"
-js-sys = "0.3.63"
-log = { workspace = true }
